Push to dev

+++ database/migrations/2019_10_05_125025_alter_table_educationalcontents_add_redirect_url.php
@@ -26,7 +26,9 @@ class AlterTableEducationalcontentsAddRedirectUrl extends Migration
     public function down()
     {
         Schema::table('educationalcontents', function (Blueprint $table) {
-            $table->dropColumn('redirectUrl');
+            if (Schema::hasColumn('educationalcontents', 'redirectUrl')) {
+                $table->dropColumn('redirectUrl');
+            }
         });
     }
 }
